refactor(davinci): split the inventory scanner under the source-length gate
tools/davinci/sourcelocation-inventory.mjs was a new 438-line file, over
the 350-line source-length budget. Scanning, member matching, and
citation resolution move to tools/davinci/lib/sourcelocation-scan.mjs
(158 lines); the entry keeps rendering and the CLI (299 lines). Output
is byte-identical - the committed artifact still passes --check.
